refactor: :zap: restructure entities and imports for improved dependency management
