+++ package.json
@@ -9,7 +9,6 @@
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "preview": "vite preview",
-    "build-gen:all": "tsc -P tsconfig.types.json && node src/entities/gen-types.js && eslint -c eslint.config.js src/entities --fix && prettier --write src/entities",
     "build:docker": "docker build -t caselab2024-team2 .",
     "start:docker": "docker run -d -p 5173:80 --rm --name caselab2024-team2 caselab2024-team2",
     "stop:docker": "docker stop caselab2024-team2",
