+++ src/app/providers/router/AppRouter.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement, ReactNode } from 'react';
 
-import { authStore } from '@/entities/auth';
+import { authStore } from '@/entities/auth/model/@x';
 import { Admin } from '@/pages/adminPage';
 import { CreateAttributePage } from '@/pages/createAttributePage';
 import { DocumentTypesPage } from '@/pages/documentPage';
