+++ src/entities/attribute/api/api-attribute.ts
@@ -1,6 +1,6 @@
 import { api } from '@/shared/http';
 
-import type { AttributeRequest, AttributeResponse } from '../index';
+import type { AttributeRequest, AttributeResponse } from '../model/@x';
 
 // получить атрибут по id
 export const getAttributeDoc = async (id: number): Promise<AttributeResponse> => {
