--- src/entities/attribute/index.ts
@@ -1,4 +1,0 @@
-export type * from '@/entities/attribute/model/types/attributeValueRequest.type';
-export type * from '@/entities/attribute/model/types/attributeValueResponse.type';
-export type * from '@/entities/attribute/model/types/attributeRequest.type';
-export type * from '@/entities/attribute/model/types/attributeResponse.type';
