+++ src/entities/attribute/model/@x/documents.ts
@@ -1,0 +1,4 @@
+//Для documents [DocumentVersionResponse]
+export type * from '../types/attributeValueResponse.type';
+//Для documents [UpdateDocumentRequest, PatchDocumentRequest, CreateDocumentRequest]
+export type * from '../types/attributeValueRequest.type';
