+++ src/entities/attribute/model/@x/index.ts
@@ -1,0 +1,6 @@
+export type * from '../types/attributeValueRequest.type';
+export type * from '../types/attributeValueResponse.type';
+export type * from '../types/attributeRequest.type';
+export type * from '../types/attributeResponse.type';
+
+export * from '../attributeStore';
