--- src/entities/auth/index.ts
@@ -1,1 +1,0 @@
-export { authStore } from './model/store';
