+++ src/entities/auth/model/@x/index.ts
@@ -1,0 +1,4 @@
+export type * from '../types/authenticationRequest.type';
+export type * from '../types/authenticationResponse.type';
+//Стор
+export { authStore } from '../store';
