+++ src/entities/auth/model/types/authenticationRequest.type.ts
