+++ src/entities/auth/model/types/authenticationResponse.type.ts
