+++ src/entities/documents/model/@x/index.ts
@@ -1,0 +1,11 @@
+//Общий публичный API
+export type * from '../types/updateDocumentRequest.type';
+export type * from '../types/documentFacadeResponse.type';
+export type * from '../types/documentPermissionResponse.type';
+export type * from '../types/documentResponse.type';
+export type * from '../types/documentVersionResponse.type';
+export type * from '../types/updateDocumentRequest.type';
+export type * from '../types/patchDocumentRequest.type';
+
+//Стор для индекса
+export type * from '../documentTypesStore';
