+++ src/entities/documents/model/@x/user.ts
@@ -1,0 +1,2 @@
+//type для user
+export type * from '../types/documentPermissionResponse.type';
