+++ src/entities/documents/model/types/documentFacadeResponse.type.ts
@@ -1,6 +1,7 @@
-import type { DocumentResponse } from '@/entities/documents/model/types/documentResponse.type.ts';
-import type { DocumentVersionResponse } from '@/entities/documents/model/types/documentVersionResponse.type.ts';
-import type { SignatureResponse } from '@/entities/signature/model/types/signatureResponse.type.ts';
+import type { DocumentResponse } from '@/entities/documents/model/@x';
+import type { DocumentVersionResponse } from '@/entities/documents/model/@x';
+import type { SignatureResponse } from '@/entities/signature/model/@x/documents';
+
 export interface DocumentFacadeResponse {
   document: DocumentResponse;
   latest_version: DocumentVersionResponse;
