--- src/entities/documents/model/types/documentTypeRequest.type.ts
@@ -1,5 +1,0 @@
-import type { DocumentTypeToAttributeRequest } from '@/entities/documents/model/types/documentTypeToAttributeRequest.type.ts';
-export interface DocumentTypeRequest {
-  name: string;
-  attributes: DocumentTypeToAttributeRequest[];
-}
