--- src/entities/documents/model/types/documentTypeResponse.type.ts
@@ -1,6 +1,0 @@
-import type { DocumentTypeToAttributeResponse } from '@/entities/documents/model/types/documentTypeToAttributeResponse.type.ts';
-export interface DocumentTypeResponse {
-  attributes: DocumentTypeToAttributeResponse[];
-  id: number;
-  name: string;
-}
