+++ src/entities/documents/model/types/updateDocumentRequest.type.ts
@@ -1,4 +1,5 @@
-import type { AttributeValueRequest } from '@/entities/attribute/model/types/attributeValueRequest.type.ts';
+import type { AttributeValueRequest } from '@/entities/attribute/model/@x/documents';
+
 export interface UpdateDocumentRequest {
   document_type_id: number;
   name: string;
