+++ src/entities/documentsType/model/@x/index.ts
@@ -1,0 +1,6 @@
+export type * from '../types/documentTypeRequest.type';
+export type * from '../types/documentTypeToAttributeRequest.type';
+export type * from '../types/documentTypeResponse.type';
+export type * from '../types/documentTypeToAttributeResponse.type';
+
+export * from '../documentTypesStore';
