+++ src/entities/documentsType/model/types/documentTypeRequest.type.ts
@@ -1,0 +1,6 @@
+import type { DocumentTypeToAttributeRequest } from './documentTypeToAttributeRequest.type.ts';
+
+export interface DocumentTypeRequest {
+  name: string;
+  attributes: DocumentTypeToAttributeRequest[];
+}
