+++ src/entities/documentsType/model/types/documentTypeResponse.type.ts
@@ -1,0 +1,7 @@
+import type { DocumentTypeToAttributeResponse } from './documentTypeToAttributeResponse.type.ts';
+
+export interface DocumentTypeResponse {
+  attributes: DocumentTypeToAttributeResponse[];
+  id: number;
+  name: string;
+}
