+++ src/entities/documentsType/model/types/documentTypeToAttributeRequest.type.ts
