+++ src/entities/documentsType/model/types/documentTypeToAttributeResponse.type.ts
