--- src/entities/signature/index.ts
@@ -1,2 +1,0 @@
-export type * from '@/entities/signature/model/types/signatureResponse.type';
-export type * from '@/entities/signature/model/types/signatureCreateRequest.type';
