+++ src/entities/signature/model/@x/documents.ts
@@ -1,0 +1,2 @@
+//Для documents [DocumentFacadeResponse]
+export type * from '../types/signatureResponse.type';
