+++ src/entities/signature/model/@x/index.ts
@@ -1,0 +1,2 @@
+export type * from '../types/signatureResponse.type';
+export type * from '../types/signatureCreateRequest.type';
