--- src/entities/user/index.ts
@@ -1,6 +1,0 @@
-export type * from '@/entities/user/model/types/userUpdateRequest.type';
-export type * from '@/entities/user/model/types/userResponse.type';
-export type * from '@/entities/user/model/types/userToDocumentResponse.type';
-export type * from '@/entities/user/model/types/registerRequest.type';
-export type * from '@/entities/user/model/types/authenticationRequest.type';
-export type * from '@/entities/user/model/types/authenticationResponse.type';
