+++ src/entities/user/model/@x/documents.ts
@@ -1,0 +1,2 @@
+//Для documents [DocumentResponse]
+export type * from '../types/userToDocumentResponse.type';
