+++ src/entities/user/model/@x/index.ts
@@ -1,0 +1,6 @@
+export type * from '../types/userUpdateRequest.type';
+export type * from '../types/userResponse.type';
+export type * from '../types/userToDocumentResponse.type';
+export type * from '../types/registerRequest.type';
+
+export * from '../userStore';
