+++ src/entities/user/model/types/userResponse.type.ts
@@ -1,5 +1,5 @@
 export interface UserResponse {
   email: string;
   display_name: string;
-  roles: string[];
+  roles: ['ADMIN' | 'USER'];
 }
