+++ src/entities/user/model/types/userToDocumentResponse.type.ts
@@ -1,4 +1,5 @@
-import type { DocumentPermissionResponse } from '@/entities/documents/model/types/documentPermissionResponse.type.ts';
+import type { DocumentPermissionResponse } from '@/entities/documents/model/@x/user';
+
 export interface UserToDocumentResponse {
   email: string;
   document_permissions: DocumentPermissionResponse[];
