--- src/entities/value.json
@@ -1,76 +1,0 @@
-{
-  "user/model/types/userToDocumentResponse": [
-    {
-      "place": "documents/model/types/documentPermissionResponse.type.ts",
-      "nameRef": "DocumentPermissionResponse"
-    }
-  ],
-  "documents/model/types/updateDocumentRequest": [
-    {
-      "place": "attribute/model/types/attributeValueRequest.type.ts",
-      "nameRef": "AttributeValueRequest"
-    }
-  ],
-  "documents/model/types/documentFacadeResponse": [
-    {
-      "place": "documents/model/types/documentResponse.type.ts",
-      "nameRef": "DocumentResponse"
-    },
-    {
-      "place": "documents/model/types/documentVersionResponse.type.ts",
-      "nameRef": "DocumentVersionResponse"
-    },
-    {
-      "place": "signature/model/types/signatureResponse.type.ts",
-      "nameRef": "SignatureResponse"
-    }
-  ],
-  "documents/model/types/documentResponse": [
-    {
-      "place": "user/model/types/userToDocumentResponse.type.ts",
-      "nameRef": "UserToDocumentResponse"
-    }
-  ],
-  "documents/model/types/documentVersionResponse": [
-    {
-      "place": "attribute/model/types/attributeValueResponse.type.ts",
-      "nameRef": "AttributeValueResponse"
-    }
-  ],
-  "documents/model/types/documentTypeRequest": [
-    {
-      "place": "documents/model/types/documentTypeToAttributeRequest.type.ts",
-      "nameRef": "DocumentTypeToAttributeRequest"
-    }
-  ],
-  "documents/model/types/documentTypeResponse": [
-    {
-      "place": "documents/model/types/documentTypeToAttributeResponse.type.ts",
-      "nameRef": "DocumentTypeToAttributeResponse"
-    }
-  ],
-  "documents/model/types/createDocumentRequest": [
-    {
-      "place": "attribute/model/types/attributeValueRequest.type.ts",
-      "nameRef": "AttributeValueRequest"
-    }
-  ],
-  "documents/model/types/patchDocumentRequest": [
-    {
-      "place": "attribute/model/types/attributeValueRequest.type.ts",
-      "nameRef": "AttributeValueRequest"
-    }
-  ],
-  "vote/model/types/voteResponse": [
-    {
-      "place": "vote/model/types/voteUserResponse.type.ts",
-      "nameRef": "VoteUserResponse"
-    }
-  ],
-  "vote/model/types/votingProcessResponse": [
-    {
-      "place": "vote/model/types/voteResponse.type.ts",
-      "nameRef": "VoteResponse"
-    }
-  ]
-}
