--- src/entities/vote/index.ts
@@ -1,5 +1,0 @@
-export type * from '@/entities/vote/model/types/votingProcessRequest.type';
-export type * from '@/entities/vote/model/types/voteResponse.type';
-export type * from '@/entities/vote/model/types/voteUserResponse.type';
-export type * from '@/entities/vote/model/types/votingProcessResponse.type';
-export type * from '@/entities/vote/model/types/voteRequest.type';
