+++ src/entities/vote/model/@x/index.ts
@@ -1,0 +1,5 @@
+export type * from '../types/votingProcessRequest.type';
+export type * from '../types/voteResponse.type';
+export type * from '../types/voteUserResponse.type';
+export type * from '../types/votingProcessResponse.type';
+export type * from '../types/voteRequest.type';
