+++ src/entities/vote/model/types/voteResponse.type.ts
@@ -1,4 +1,5 @@
-import type { VoteUserResponse } from '@/entities/vote/model/types/voteUserResponse.type.ts';
+import type { VoteUserResponse } from './voteUserResponse.type.ts';
+
 export interface VoteResponse {
   status: string;
   applicationUser: VoteUserResponse;
