+++ src/pages/documentPage/ui/DocumentFormModalWindow.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 
-import { attributesStore } from '@/entities/attribute/model/attributeStore';
+import { attributesStore } from '@/entities/attribute/model/@x';
 import { Box, Button, Container, Modal, TextField, Typography } from '@mui/material';
 import { Form, Formik } from 'formik';
 import { observer } from 'mobx-react-lite';
